+++ src/test/java/domain/ladder/HeightTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import domain.ladder.Height;
 import org.junit.jupiter.api.DisplayName;
