+++ src/test/java/domain/ladder/LadderTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import domain.ladder.Height;
 import domain.ladder.Ladder;
