+++ src/test/java/domain/ladder/LineTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import domain.ladder.Line;
 import domain.ladder.stick.Stick;
