+++ src/test/java/domain/result/ResultsTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 import domain.result.Result;
 import domain.result.Results;
